little warning fixes and workarounds for various platforms

+++ fmt/mus.c
@@ -48,8 +48,9 @@ int fmt_mus_read_info(dmoz_file_t *file, const uint8_t *data, size_t length)
 {
         struct mus_header *hdr = (struct mus_header *) data;
 
+        /* cast necessary for big-endian systems */
         if (!(length > sizeof(*hdr) && memcmp(hdr->id, "MUS\x1a", 4) == 0
-              && bswapLE16(hdr->scorestart) + bswapLE16(hdr->scorelen) <= length))
+              && (size_t) (bswapLE16(hdr->scorestart) + bswapLE16(hdr->scorelen)) <= length))
                 return 0;
 
         file->description = "Doom Music File";
